requirements.txt
streamlit
pandas
python-docx
gspread
google-auth-oauthlib
google-api-python-client